// web/src/api/db/types.ts
export enum EMainRoutes {
  User = '/user',
  Home = '/home',
  Authentication = '/authentication',
}

export enum ESecondaryRoutes {
  GetMe = '/get-me',
  PublishPost = '/',
  SignUp = '/sign-up',
  SignIn = '/sign-in',
  GetPosts = '/posts',
  IntentResetPassword = '/intent-reset-password',
  SubmitResetPassword = '/submit-reset-password',
}